[REFACTOR][SCRIPT] Lift TVMScript entry point into script/printer/printer.h
Add a new public header `include/tvm/script/printer/printer.h` that
declares the free function `tvm::Script(node, optional_config)` plus
the dispatch vtable used by per-dialect printers. `PrinterConfig` and
related dataclass helpers stay in `config.h`; `printer.h` is the
entry-point header that callers include to invoke printing.

The `TVMScriptPrinter::Script(...)` static method body moves to the
free function `tvm::Script(...)` in `src/script/printer/script_printer.cc`.
`TVMScriptPrinter` class and `TVM_REGISTER_SCRIPT_AS_REPR` macro move
from `config.h` to `printer.h`.

Existing direct callers in s_tir/schedule/error.cc, script/printer/config.cc,
and script/printer/utils.h are rewritten to the new entry point.
